Hashing refactoring
New script called full_command_hash.py in which the user provides the
machine_hash and guffi_command hash to generate a combined hash that is
stored in the full_hash table. This change now means that the user is
only required to provide the combined hash as an argument to
extraction.py rather than the machine_hash and guffi_hash.